test_project/requirements.txt
django-fias=2.0.1
django_select2>=5.3.0
rarfile
six
lxml
unrar
zeep>=0.17.0
dbfread>=2.0.0
progress
mysqlclient != 1.3.8
